Prelimenary support playground of HTTP basic authentication and HTTP SSO
Includes improvements to URL handling (by using URI) and a reorganization of
Handler. A logging.properties file was added to aid to squenching log levels.
The Handlers now support the HEAD request, and respond appropriately to HTTP
request types they do not support.

It would probably be a good idea to swap to using java.net.HttpURLConnection
for defines of HTTP status codes instead of using the plain error code integer.