Reorganize classes to make packages more helpful
The separate-package-for-each-Adaptor rule has begun causing more noise
than is helpful, especially since each Adaptor is only one or two
classes in size. The JavaDoc is especially fragmented due to it as well.
Also, there are several transforms and other classes in adaptorlib that
are for very specific use-cases, causing clutter.

The reorganized form has three packages: adaptorlib,
adaptorlib.examples, and adaptorlib.prebuilt. If an Adaptor grows too
large, then a sub-package can be created. All Adaptors and Transforms
should go in either examples or prebuilt, depending on their expected
usage. The AutoUnzipAdaptor remains in adaptorlib, but it will likely
move in the future, depending on where we take it.

This reorganization is also intended to aid packaging the core library
with as few 3rd-party dependencies as possible.