Implement HTTPS communication and Authn via SecMgr
HTTPS and Authn are enabled by the 'server.secure' configuration
variable. When enabled, you must generate and exchange keys with the
GSA. Both the GSA and the adaptor require generated keys because the
GSA's default certificate is for the wrong domain. You must also copy
the jars from lib/endorsed into your JRE's lib/endorsed directory.

The Security Manager is assumed to be on the GSA.