Fork once for all JUnit test cases
This is instead of forking for each test case. Note that the reloading
property is enabled by default, so the classloader is instantiated for
each test case (so static fields within our code are re-instantiated).
The tests can get another 2x speedup by not forking at all, but 1) the
performance win is a fixed amount total - not per test and 2) the code
coverage instrumentation is most easily recorded by exiting the JVM.