Limit Dashboard to GSA Administrators
Also only create sessions when required. Users are prompted with a form
for authentication. When they submit the form, we try logging in to the
GSA's administrative APIs with their credentials. Depending on if it
succeeds, they are permitted to view the Dashboard. If success, then the
positive result is stored in their session.