Use cron-style scheduler instead of replacable one
The previous scheduler allowed providing different implementations for
the Iterable<Date>. However, it was determined that the cron format
should be good enough for all reasonable schedules. Thus, the previous
scheduler has been removed and a cron one added without the ability to
swap to a different scheduler. The end user can now specify the push
schedule with the adaptor.fullListingSchedule configuration key.