Large documents now circumvent transform pipeline
This allows us to keep memory usage under control. The user can choose
the upper limit they want to allow going through the transform pipeline.
The code buffers a document until the moment it exceeds the maximum
transform size, at which point it begins streaming to the client and
avoids the buffer.