Check common name of client certificates
We now check the common name of the subject of client certificates when
considering if a HTTP client should be provided full access to
documents. There is a whitelist of valid common names to provide full
access. Previously we expected only the GSA's certificate to be in the
truststore, but now we allow other, unrelated certificates to be in the
same store.

Since we expect the common names to be hostnames (although this is not a
requirement), we use the same list as the basis for IP-based security
when secure mode is disabled.