Make TransformPipeline immutable and not be a List
TransformPipeline no longer extends AbstractList and is provided the
transform list during construction. It is still possible to change the
behavior of a pipeline after creation by modifying a DocumentTransform
instance since they can be mutable, but that would be looked down upon.

Although making the TransformPipeline immutable is nice, the primary
benefit is reducing the amount of API we expose in addition to getting
rid of the awkward implementation of List.