Improve empty ACL support
This includes documentation for how it is handled, as well as support
for both handling it the way the GSA appears to handle it vs the way the
GSA actually handles it. This distinction will be important once
Adaptors have a way to inform the GSA that something is private.