Add support for X-Gsa-Serve-Security
This allows us to support authz mechanisms other than ACLs, because we
are now easily able to inform the GSA that a document is secure. In
addition, the header allows us to mark documents as public, since GSAs
can now be configured to mark documents as secure when the server
requests a client certificate. This feature requires GSA 7.0 or later.

An alternative to X-Gsa-Serve-Security is requiring HTTP Basic
Authentication, but this requires the GSA be properly configured and
only is capable of marking secure documents.