Send ACL with fake deny rule to GSA for empty ACL
The empty ACL cannot normally be distinguished from a null ACL on the
GSA, because there is actually no way to encode a completely empty ACL.
In order to simplify the API and prevent unexpected consequences, we
are now sending a replacement ACL that has a trash deny rule so that
the GSA can detect the ACL's existance, but will still resolve to
INDETERMINATE for all normal cases.

Since the empty ACL no longer means "public," the emptyImpliesPublic
flag has been removed. EmptyImpliesPublic is essentially always false
now.