Make uses of CommandStreamParser consistant
Previously, CommandStreamParser required copying various pieces of data
to the Response by the user of the parser. Now, the parser interacts
directly with the Response, which allows all consumers of the API to
immediately recieve benefits of additions to the file format without any
changes on their end.

In addition, using the Response directly allows streaming data from the
InputStream to the OutputStream, without buffering the entire document
contents first, and batching records to the GSA in the lister
occasionally.