Remove unnecessary public API from HttpExchanges
Respond(), respondToHead(), and startResponse() are all too specific and
tuned to our own usage to expose to others. What consumers of this API
would really need is a cannedRespond(), but that will take some effort
to produce a reasonable API (due to localization).

Since we are removing the response methods, we had to improve
enableCompressionIfSupported to not rely on them or stop exposing the
method. This commit choses not to rely on them.