Revert "send namespaced acls in http headers"
This reverts commit 0e933f493c9a28534f0f43a98468990956f70327.

Due to miscommunication, this got committed before full review. It's
also questionable if we want to commit it at this time since some of its
dependencies on the GSA are not yet in place. It's possible this change
is reintroduced at a later time exactly as it is now, but we need time
for discussion.