Allow specifying a prefix for HTTP context paths
Config.getServerBaseUri(DocId) was removed because it was not adequately
exposed to be heavily used by customers, and because its use case is now
intended to be handled by being behind a load-balancing reverse proxy.
In addition, its removal simplified the interface of DocIdCodec, now
that it must also be informed of the context prefix.