Add batching for group definition feeds
This change is to address concerns over memory usage. Since the entire
feed is buffered in memory, batching the feeds allows us to reduce the
size of the block of memory allocated (which reduces the effects of
fragmentation) and reduces the amount of memory required at any point in
time.