Fix calling watchdog from different thread
Previously it was assumed that Response would only be accessed on the
primary thread for a request. However, our own builtin Command Line
Adaptor actually used Response from a different thread. With this change
you can now use Response on a different thread than getDocContent(),
although the object is still not thread-safe.