Add a diagnostic DownloadDumpHandler class
Create a new Handler class (DownloadDumpHandler) that generates a new
Zip file when it matches the URL /diagnostics-zip-dump.  Currently, that
file contains all the system logs as well as a thread dump, but future
versions will also include a copy of the configuration as well as other
diagnostics.

New file src/com/google/enterprise/adaptor/DownloadDumpHandler.java
contains most of the logic; the only changes to the other files are to
register the new Handler (in Dashboard.java) and to provide a link
that generates the Zip into resources/.../resources/index.html .  In addition,
new functionality for the MockFile test class (implemented by ejona@) is
included.  The DownloadDumpHandlerTest class uses this new functionality.

Unit test suite for DownloadDumpHandler has 98% code coverage; the only line not
covered is the one method overriden in the test class (to avoid real file I/O).

Made only the one-argument constructor (with the feedname) public; the other is
package-private (as it's only for testing the class).