LIB Make AdaptorTemplate.jar with manifest for testing service
This adds a build target for constructing a stand-alone AdaptorTemplate.jar for
use in testing. The jar requires only a matching adaptor-[version]-withlib.jar
and a adaptor-config.properties in the same directory as AdaptorTemplate.jar.

It can be run with the command 'java -jar AdaptorTemplate.jar'

Code Review: http://codereview.appspot.com/73830043