Fix b/15187210: Better log message when dashboard login fails
This change will include the GSA admin URL in the log messages when
failing to connect the the GSA admin port for authentication.
Better, more descriptive log messages were crafted for common
GSA admin communication issues that can be fixed with proper
configuration of gsa.hostname, gsa.admin.hostname, and SSL certs.

This only fixes messages logged in the log file, and not the error
message that is displayed on the dashboard login page.

Code Review: http://codereview.appspot.com/110200046