Do not send ACLs and Groups to GSA if markAllDocsAsPublic is true.
If markAllDocsAsPublic is true, feeding ACLs and Groups serves no
purpose.  This modifies DocIdSender to return immediately if
pushNamedResources() or pushGroupDefinitions() are called when
markAllDocsAsPublic is true.

When I added tests for this, I found some of the tests for
pushing named resources and groups normally were deficient,
so I beefed them up a bit.

Code Review: http://codereview.appspot.com/108400046