Log product version on startup.
This logs the name and version of the Adaptor
and the Adaptor Library on startup.  This was
already done for Dashboard.  However this change
makes Dashboard a bit smarter displaying the
Adaptor name, using the product title from the
Manifest if it is available, and the simple
class name, if not.

Code Review: http://codereview.appspot.com/140260043