Compile Adaptor Library with Java 6 BootClasspath
This change builds the Adaptor library, examples, and tests with
a Java 6 bootclasspath to avoid accidentally using classes and
methods introduced in Java 7 and 8.

The bootclasspath is read from a new build.properties file which
contains developer environment configuration properties.  In this
case, the location of a JDK 6 installation.

Lastly, this change updates one of the experimental classes, as it
used a now-deprectated method.

Code Review: http://codereview.appspot.com/234820043