Adds configuration option to disable full and incremental listing.
This change adds a new configuration option,
adaptor.disableFullAndIncrementalListing.  If true, the plexi
library refrains from calling the adaptor's getDocIds() and
getModifiedDocIds() methods.

Also, if true, the values of the adaptor.pushDocIdsOnStartup,
adaptor.fullListingSchedule, and adaptor.incrementalPollPeriodSecs
properties are ignored; and the full and incremental feed buttons
on the dashboard are disabled.

The AsyncDocIdPusher is still active, but adaptors should restrict
its use to Adaptor.getDocContent().

This change also subtly alters the scheduling of calls to
getModifiedDocIds(). Previously, the exectutor schedule
had no initial delay. This resulted and getDocIds() and
getModifiedDocIds() being called simultaneously on startup.
This change sets the initial delay to the value of
adaptor.incrementalPollPeriodSecs if adaptor.pushDocIdsOnStartup
is true (the default), and zero if adaptor.pushDocIdsOnStartup
is false.

Code Review:  http://codereview.appspot.com/260700043