Add a forced transmission decision that should not be overridden
Graph traversal connectors break when folders are skipped,
so they need a way to override filters that may attemt to
skip these folders. This change adds a Forced Transmission
Decision that overrides any Transmission Decision set by
the filters. As a convenience, our built-in filters defer
to a forced decision and return immediately.