Expose a recording Response fake for unit testing.
Rename WrapperAdaptor.GetContentsResponse to RecordingResponse, make
it public, and move it to the testing package. This class was also
copied to the SharePoint connector, and very similar classes written
for the file system, Documentum, and OpenText connectors. The library
and SharePoint versions were the most complete; the others were
missing a handful of the flag methods (e.g., setNoFollow).

Only minor changes were made, including some merging of the way the
output stream, metadata, and anchors were handled in the various
implementations. The fields and methods were rearranged to match the
Response source file.