Add RegexDecisionFilter MetadataTransform (#29)
RegexDecisionFilter makes a TransmissionDecision using regular expression matching on metadata and/or params values. It is basically SkipDocumentFilter with a configurable decision. RegexDecisionFilter takes the following configuration parameters:

key: the metadata or parameter key whose values will be matched against pattern.

keyset: location of key, "metadata" or "params"

pattern: a regular expression used to match values. if unspecified, anything will match (effectively tests presence/absence of key).

when: if "found" the decision is applied if a match is found. if "not-found" the decision is applied if no match is found.

decision: one of TransmissionDecision string values: "as-is", "do-not-index", "do-not-index-content".