Expose a recording DocIdPusher fake for unit testing. (#54)
Part 1: Merge the various implementations.

Rename AccumulatingDocIdPusher to RecordingDocIdPusher, make it
public, and move it to the testing package.This class was also copied
to five of the connectors, where they slightly diverged. The
differences were merged into this copy. Make AbstractDocIdPusher
public. Only the library's version extended AbstractDocIdPusher, but
they all contained exact copies of AbstractDocIdPusher's methods.
The methods were rearranged for the getters to follow the setters.

Semantic changes and javadoc will follow in later changes. This one
contains the necessary functional changes, and comes in just over
git's default similarity index of 50% for recognizing renamed files.