Undo changes to GsaFeedFileArchiver and GsaFeedFileSender.
Revert the changes to these files from commit 48c91fb, and in
particular restore them to package-private access so that they do
not appear in the published javadoc. They were made public as
part of the send2gsa effort, but the changes were not reverted in
commit f448c2f.

Keep a TODO about useCompression.