Allow empty group definitions map to be pushed. (#76)
The connector should be able to nuke all the group definitions by sending an empty full feed, but that was short circuited by the check for defs.isEmpty() in the beginning of pushGroupDefinitionsInternal(). This bypassed the FULL group feed logic at the end of the method that sends an empty full feed to the GSA that will delete previous group definitions. So attempting to delete all the groups fails.

This change removes the short circuit check at the beginning of pushGroupDefinitionsInternal(), allowing the empty group definitions map to be processed normally.  The while loop will be skipped, so a feed based on the map will never be sent to the GSA. If the feedType was INCREMENTAL, we return and the effect from the outside looks the same as the short circuit. If the feedType was FULL, then the empty full feed will be sent at the end of pushGroupDefinitionsInternal(), removing all group definitions from the GSA. Minor changes were made to code that assumed the map was not empty.