Switch from asList to ImmutableList.of in DodIdSenderTest.
Part of removing the suppression of rawtypes warnings.

This change started by avoiding the creation of explicit arrays to
pass to the Arrays.asList varargs method, to reduce clutter and make
it easier to reason about the many nested collections in this class.

Two calls to asList use parameterized types, which results in a
warning about unchecked generic array creation. ImmutableList.of
happens to have overloads with multiple arguments rather than arrays.

This class already used both asList (25%) and ImmutableList.of (75%).
Since I am already editing nearly every call to asList to remove the
explicit arrays, switching the class to consistently use
ImmutableList.of seems reasonable.