Fix PrebuiltTranform.replaceInParams empty and trimmed values
The configurations for replace string, pattern, and replacement args
are intentionaly not trimmed, allowing patterns, etc, to have leading
and trailing whitespace. Unfortunately PrebuiltTranform.replaceInParams()
would trim the original values read (and perform emptyToNull conversion).
This could cause the value to not match the string or pattern and failed
on empty values.

This change causes replaceInParams() to use the unmodified original values
for the parameters. It also adds more extensive tests for whole value
replacements; empty Strings for string, pattern, replacement, and values;
and untrimmed Strings for string, pattern, replacement, and values.